requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
msgpack==1.0.7
Pillow==10.1.0
psycopg2==2.9.9
py-algorand-sdk==2.5.0
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-decouple==3.8
pytz==2023.3.post1
ruff==0.1.4
sqlparse==0.4.4
typing-extensions==4.8.0
